refactor: rename to pqrs.store.ts
